+++ .gitignore
@@ -1,4 +1,4 @@
 out/
 node_modules/
-/test/*.js
+/test/*-out/
 /.turbo
