+++ turbo.json
@@ -18,14 +18,14 @@
       "dependsOn": ["test:babel:build"]
     },
     "test:babel:build": {
-      "outputs": ["test/babel.js"],
+      "outputs": ["test/js-out/babel.js"],
       "dependsOn": ["build"]
     },
     "test:ts": {
       "dependsOn": ["test:ts:build"]
     },
     "test:ts:build": {
-      "outputs": ["test/typescript.js"],
+      "outputs": ["test/ts-out/typescript.js"],
       "dependsOn": ["build"]
     }
   }
